+++ R/read_plink.R
@@ -1,18 +1,19 @@
 #' Read genotype and sample data in a Plink BED/BIM/FAM file set.
 #'
-#' This function reads a genotype matrix (`X`) and its associated locus (`bim`) and individual (`fam`) data tables in the three Plink files in BED, BIM, and FAM formats, respectively.
+#' This function reads a genotype matrix (`X`, encoded as reference allele dosages) and its associated locus (`bim`) and individual (`fam`) data tables in the three Plink files in BED, BIM, and FAM formats, respectively.
 #' All inputs must exist or an error is thrown.
 #' This function is a wrapper around the more basic functions
 #' [read_bed()],
 #' [read_bim()],
-#' [read_fam()].
+#' [read_fam()],
+#' which simplifies data parsing and additionally better guarantees data integrity.
 #' Below suppose there are `m` loci and `n` individuals.
 #' 
 #' @param file Input file path, without extensions (each of .bed, .bim, .fam extensions will be added automatically as needed).
 #' Alternatively, input file path may have .bed extension (but not .bim, .fam, or other extensions).
 #' @param verbose If `TRUE` (default), function reports the paths of the files being read (after autocompleting the extensions).
 #'
-#' @return A named list with items in this order: `X` (genotype matrix), `bim` (tibble), `fam` (tibble).
+#' @return A named list with items in this order: `X` (genotype matrix, see description in return value of [read_bed()]), `bim` (tibble, see [read_bim()]), `fam` (tibble, see [read_fam()]).
 #' `X` has row and column names corresponding to the `id` values of the `bim` and `fam` tibbles.
 #'
 #' @examples
@@ -51,8 +52,11 @@
 #'
 #' @seealso
 #' [read_bed()],
-#' [read_bim()],
-#' [read_fam()].
+#' [read_bim()], and
+#' [read_fam()]
+#' for individual parsers of each input table, including a description of each object returned.
+#' 
+#' [geno_to_char()] for translating numerical genotypes into more human-readable character encodings.
 #' 
 #' Plink BED/BIM/FAM format reference:
 #' <https://www.cog-genomics.org/plink/1.9/formats>