+++ tests/testthat/test-write_bed.R
@@ -311,3 +311,82 @@ test_that("write_plink with `append = TRUE` works", {
     expect_silent( delete_files_plink(fo) )
 })
 
+test_that( "geno_to_char works", {
+    # use data where plink1 told us what the answer was
+    name <- 'dummy-4-10-0.1'
+    data <- read_plink( name )
+    X <- data$X
+    bim <- data$bim
+    m <- nrow( X )
+    n <- ncol( X )
+    
+    # need to primitively parse corresponding ped output
+    # (sorry this is tedious)
+    name_ped <- paste0( name, '.ped' )
+    Y_exp <- readr::read_table( name_ped, col_names = FALSE )
+    # remove FAM columns (first 6)
+    # also transpose
+    Y_exp <- t( Y_exp[ , -(1:6) ] )
+    # and collapse the way we have it
+    for ( i in 1 : m ) {
+        # indexes on Y matrix
+        # first row
+        j <- 2 * i - 1
+        # get two rows at the time
+        # store back in final row
+        Y_exp[ i, ] <- paste0( Y_exp[ j, ], '/', Y_exp[ j + 1, ] )
+    }
+    # shrink matrix now
+    Y_exp <- Y_exp[ 1:m , ]
+    # cases are coded as 0/0 are NA, replace now:
+    Y_exp[ Y_exp == '0/0' ] <- NA
+    # copy correct names (missing in this data)
+    dimnames( Y_exp ) <- dimnames( X )
+    # for simplicity, flip heterozygotes so they're all A/B instead of some being B/A
+    Y_exp[ Y_exp == 'B/A' ] <- 'A/B'
+
+    # START TESTING!
+
+    # errors due to missing required arguments
+    expect_error( geno_to_char( X = X ) )
+    expect_error( geno_to_char( bim = bim ) )
+    # disagreeing dimensions
+    expect_error( geno_to_char( X, bim[ -1, ] ) )
+    # slashes in allele names
+    bim_bad <- bim
+    i <- sample( m, 1 )
+    bim_bad$ref[ i ] <- 'a/b'
+    expect_error( geno_to_char( X, bim_bad ) )
+    bim_bad <- bim
+    i <- sample( m, 1 )
+    bim_bad$alt[ i ] <- 'a/b'
+    expect_error( geno_to_char( X, bim_bad ) )
+    # same ref/alt at a locus
+    bim_bad <- bim
+    i <- sample( m, 1 )
+    bim_bad$alt[ i ] <- bim_bad$ref[ i ]
+    expect_error( geno_to_char( X, bim_bad ) )
+    # invalid values in X
+    X_bad <- X
+    i <- sample( m*n, 1 )
+    X_bad[ i ] <- -1
+    expect_error( geno_to_char( X_bad, bim ) )
+    X_bad <- X
+    i <- sample( m*n, 1 )
+    X_bad[ i ] <- 0.5
+    expect_error( geno_to_char( X_bad, bim ) )
+    X_bad <- X
+    i <- sample( m*n, 1 )
+    X_bad[ i ] <- 3
+    expect_error( geno_to_char( X_bad, bim ) )
+    
+    # now a succesful run
+    expect_silent(
+        Y_obs <- geno_to_char(X, bim)
+    )
+    # allele order doesn't matter, so normalize it here as we did with the expected matrix
+    Y_obs[ Y_obs == 'B/A' ] <- 'A/B'
+    # make sure we recovered the desired data!
+    expect_equal( Y_obs, Y_exp )
+})
+